packages/serialization/text/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "microsoft-kiota-serialization-text"
version = "0.0.4"
description = "Core abstractions for kiota generated libraries in Python"
authors = ["Microsoft <graphtooling+python@microsoft.com>"]
license = "MIT"
readme = "README.md"
keywords = ["kiota", "openAPI", "Microsoft", "Graph"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
homepage = "https://github.com/microsoft/kiota#readme"
repository = "https://github.com/microsoft/kiota-python"
documentation = "https://microsoft.github.io/kiota/"
packages = [{include = "kiota_serialization_text"}]	

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
microsoft-kiota-abstractions = ">=1.0.0,<2.0.0"
python-dateutil = "2.8.2"

[tool.poetry.group.dev.dependencies]
yapf = "^0.40.2"
isort = "^5.13.2"
pylint = "^3.2.7"
mypy = "^1.11.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[tool.mypy]
warn_unused_configs = true
files = "kiota_serialization_text"
ignore_missing_imports = true

[tool.yapf]
based_on_style = "pep8"
dedent_closing_brackets = true
each_dict_entry_on_separate_line = true
column_limit = 100

[tool.isort]
profile = "hug"